requirements.txt
requests
beautifulsoup4
selenium
python-dotenv